refactor: upload 라우터 이동
upload 라우터를 지우고 각 리소스에 해당하는 라우터로 이동
