+++ src/routes/user.ts
@@ -1,8 +1,24 @@
 import { NextFunction, Request, Response, Router } from 'express';
 import createError from 'http-errors';
+import { IS3Request, profileUpload } from '../config/multer';
 import UserController from '../controllers/user';
+import { isLoggedIn } from '../middlewares/auth';
 
 const userRouter = Router();
+
+const uploadErrorMessage = '프로필 이미지 업로드에 실패했습니다. 파일 크기(1MB 미만)나 네트워크 회선을 점검해주세요.';
+
+function profileUploadMiddleware(req, res, next) {
+  profileUpload.single('profileimg')(req, res, err => {
+    if (err) {
+      createError(500);
+      req.flash('flashMessage', uploadErrorMessage);
+      res.redirect('/');
+    } else {
+      next();
+    }
+  });
+}
 
 userRouter.get('/:id', async (req: Request, res: Response, next: NextFunction) => {
   try {
@@ -44,6 +60,20 @@ userRouter.put('/:id', async (req: Request, res: Response, next) => {
   } catch (error) {
     createError(500);
     next(error);
+  }
+});
+
+userRouter.post('/profile', isLoggedIn, profileUploadMiddleware, async (req: IS3Request, res: Response, next: NextFunction) => {
+  try {
+    await UserController.PatchUserById({
+        _id: req.user._id,
+        profileImageUrl: req.file.location,
+    });
+    return res.redirect('/');
+  } catch (error) {
+    createError(500);
+    req.flash('flashMessage', uploadErrorMessage);
+    res.redirect('/');
   }
 });
